+++ lib/tagged/status.ex
@@ -2,7 +2,8 @@ defmodule Tagged.Status do
   @moduledoc ~S"""
   Resoning in terms of the status of a result.
 
-      iex> use Tagged.Status
+      iex> require Tagged.Status
+      iex> import Tagged.Status, only: [ok: 1]
       iex> ok(:computer)
       {:ok, :computer}
       iex> with ok(it) <- Keyword.fetch([a: "bacon"], :a), do: "Chunky #{it}!"